pyproject.toml
[project]
name = "meal-mate"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "pyTelegramBotAPI==4.22.1",
    "APScheduler==3.10.4",
    "openai>=1.30.0",
    "recipe-scrapers==14.55.0",
    "python-dotenv==1.0.1",
    "pytz==2024.2",
    "requests==2.31.0",
]
